완전체

+++ src/main/java/com/oasys/common/member/vo/MemberVO.java
@@ -16,8 +16,8 @@ public class MemberVO extends CommonVO{
 	private String m_phone;	//핸드폰 번호
 	private String m_email;	//이메일
 	private String m_birth;	//생년월일
-	private int m_total;	//누적금액
-	private int m_ytotal;	//연간누적금액
+	private double m_total;	//누적금액
+	private double m_ytotal;	//연간누적금액
 	private String m_joindate;	//회원 가입일
 	private String m_gender;	//성별
 	private String m_type;	//타입(회원 or 관리자)
